// lib/api/generated/models/models-register-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ModelsRegisterRequest {
    'email': string;
    'first_name': string;
    'last_name': string;
    'middle_name'?: string;
    'password': string;
    'user_type': ModelsRegisterRequestUserTypeEnum;
    /**
     * Made optional
     */
    'username'?: string;
    'verify_password': string;
}

export const ModelsRegisterRequestUserTypeEnum = {
    Creator: 'creator',
    BrandUser: 'brand_user'
} as const;

export type ModelsRegisterRequestUserTypeEnum = typeof ModelsRegisterRequestUserTypeEnum[keyof typeof ModelsRegisterRequestUserTypeEnum];